+++ lib/include/QGeoView/QGVUtils.h
@@ -18,28 +18,11 @@
 
 #pragma once
 
-#include "QGVGlobal.h"
+#include <QGVGlobal.h>
 
-#include <QGraphicsItem>
+namespace QGV {
 
-class QGVDrawItem;
+QGV_LIB_DECL double metersToDistance(const double meters, const DistanceUnits unit);
+QGV_LIB_DECL QString unitToString(const DistanceUnits unit);
 
-class QGV_LIB_DECL QGVMapQGItem : public QGraphicsItem
-{
-public:
-    explicit QGVMapQGItem(QGVDrawItem* geoObject);
-
-    static QGVDrawItem* geoObjectFromQGItem(QGraphicsItem* item);
-
-    void resetGeometry();
-
-private:
-    QRectF boundingRect() const override final;
-    void paint(QPainter* painter, const QStyleOptionGraphicsItem* option, QWidget* widget = 0) override final;
-    QPainterPath shape() const override final;
-    void hoverEnterEvent(QGraphicsSceneHoverEvent* event) override final;
-    void hoverLeaveEvent(QGraphicsSceneHoverEvent* event) override final;
-
-private:
-    QGVDrawItem* mGeoObject;
-};
+} // namespace QGV