+++ lib/src/QGVDrawItem.cpp
@@ -1,6 +1,6 @@
 /***************************************************************************
  * QGeoView is a Qt / C ++ widget for visualizing geographic data.
- * Copyright (C) 2018-2020 Andrey Yaroshenko.
+ * Copyright (C) 2018-2023 Andrey Yaroshenko.
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as published by
@@ -25,6 +25,7 @@ double highlightScale = 1.15;
 }
 
 QGVDrawItem::QGVDrawItem()
+    : mDirty{ false }
 {}
 
 void QGVDrawItem::setFlags(QGV::ItemFlags flags)
@@ -33,7 +34,6 @@ void QGVDrawItem::setFlags(QGV::ItemFlags flags)
         mFlags = flags;
         projOnFlags();
         refresh();
-        repaint();
     }
 }
 
@@ -94,6 +94,10 @@ void QGVDrawItem::refresh()
     mQGDrawItem->setOpacity(effectiveOpacity());
     mQGDrawItem->setZValue(effectiveZValue());
     mQGDrawItem->setAcceptHoverEvents(isFlag(QGV::ItemFlag::Highlightable));
+    mQGDrawItem->update();
+
+    mDirty = false;
+
     if (QGV::isDrawDebug()) {
         setProperty("updateCount", property("updateCount").toInt() + 1);
     }
@@ -101,7 +105,13 @@ void QGVDrawItem::refresh()
 
 void QGVDrawItem::repaint()
 {
-    if (!mQGDrawItem.isNull()) {
+    if (mQGDrawItem.isNull()) {
+        return;
+    }
+
+    if (mDirty) {
+        refresh();
+    } else {
         mQGDrawItem->update();
     }
 }
@@ -110,6 +120,11 @@ void QGVDrawItem::resetBoundary()
 {
     if (!mQGDrawItem.isNull()) {
         mQGDrawItem->resetGeometry();
+    }
+
+    if (isFlag(QGV::ItemFlag::Transformed) || isFlag(QGV::ItemFlag::Highlighted) ||
+        isFlag(QGV::ItemFlag::IgnoreScale) || isFlag(QGV::ItemFlag::IgnoreAzimuth)) {
+        mDirty = true;
     }
 }
 
@@ -164,6 +179,15 @@ void QGVDrawItem::projOnMouseDoubleClick(const QPointF& projPos)
     }
 }
 
+void QGVDrawItem::projOnObjectStartMove(const QPointF& /*projPos*/)
+{}
+
+void QGVDrawItem::projOnObjectStopMove(const QPointF& /*projPos*/)
+{}
+
+void QGVDrawItem::projOnObjectMovePos(const QPointF& /*projPos*/)
+{}
+
 void QGVDrawItem::onProjection(QGVMap* geoMap)
 {
     QGVItem::onProjection(geoMap);
@@ -188,14 +212,12 @@ void QGVDrawItem::onCamera(const QGVCameraState& oldState, const QGVCameraState&
         return;
     }
     refresh();
-    repaint();
 }
 
 void QGVDrawItem::onUpdate()
 {
     QGVItem::onUpdate();
     refresh();
-    repaint();
 }
 
 void QGVDrawItem::onClean()